pyproject.toml
[tool.poetry]
name = "textrankai"
version = "0.1.0"
description = ""
authors = ["DESKTOP-E1D17BN\\ysh <ysh0714@naver.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "4.0"
django-ninja = "0.16.1"
scikit-learn = "^1.0.2"
konlpy = "^0.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
